[grammar] implement Option element to control AST names
