+++ test/parser_equivalence_test.py
@@ -1,11 +1,11 @@
 
-import importlib
 from pathlib import Path
 
 import pytest  # noqa
 
 from tatsu.exceptions import FailedParse
-from tatsu.tool import compile, gencode
+from tatsu.util.testing import generate_and_load_parser
+from tatsu.tool import compile
 
 INPUT = """
     1d3
@@ -39,27 +39,6 @@ GRAMMAR = """
         | int
         ;
 """
-
-
-def generate_and_load_parser(name, grammar):
-    init_filename = Path('./tmp/__init__')
-    init_filename.touch(exist_ok=True)
-
-    parser = gencode(name='Test', grammar=grammar)
-    parser_filename = Path(f'./tmp/{name}.py')
-    with open(parser_filename, 'wt') as f:
-        f.write(parser)
-    try:
-        importlib.invalidate_caches()
-        module = importlib.import_module(f'tmp.{name}', 'tmp')
-        importlib.reload(module)
-        try:
-            return module.UnknownParser()  # noqa
-        except (AttributeError, ImportError):
-            return module.TestParser()  # noqa
-    finally:
-        pass
-        # parser_filename.unlink()
 
 
 def test_model_parse():
